refactor(权限): 增加增删查改接口

+++ src/main/java/com/fruitbasket/orange/module/rbac/pojo/query/RoleAddQuery.java
@@ -34,7 +34,7 @@ public class RoleAddQuery {
     /**
      * 排序值
      */
-    private Byte sortValue;
+    private Integer sortValue;
 
     /**
      * 备注
